decider/requirements.txt
google-adk[a2a]
google-cloud-aiplatform
vertexai
google-genai
uvicorn[standard]
fastapi